refactor: use direct parameters for space/resolution in MaskData
Replace metadata-based storage with direct attributes for cleaner API.
All MaskData instance creation now requires explicit space and resolution
parameters instead of relying on metadata dictionary.

Changes:
- MaskData stores space/resolution as _space/_resolution attributes
- Properties expose space/resolution (backward compatible)
- Fixed all MaskData() calls: copy(), add_result(), add_provenance()
- Updated BaseAnalysis.run() to pass space/resolution parameters
- Fixed transformation methods in network mapping analyses
- Updated contract tests for new error messages

Breaking change: MaskData() now requires space and resolution as
direct parameters. No backward compatibility maintained per approved
refactoring plan.

+++ tests/unit/test_parcellation_module_refactoring.py
@@ -122,7 +122,7 @@ class TestParcellationFunctionSignatures:
         from lacuna.assets.parcellations import load_parcellation, register_parcellation
 
         # Create minimal test parcellation
-        data = np.arange(1, 28).reshape(3, 3, 3)
+        data = np.arange(1, 28, dtype=np.uint8).reshape(3, 3, 3)
         img = nib.Nifti1Image(data, affine=np.eye(4))
         path = tmp_path / "test.nii.gz"
         nib.save(img, path)
